stream: limit unix read/write to 2GB at a time
On some platforms, notably OSX, read/readv/write/writev all return
an EINVAL if the requested number of bytes is greater than INT32_MAX.

We therefore limit each such call to 2GB on all platforms. On
platforms where this is not an issue, the overhead of an extra syscall
for every 2GB of read/write is negligible.